fixes issue #510
assign a random password if none is supplied.  this is what the API spec has always said would happen, but for some reason, it wasn't happening

now it does.  note that if you have any kind of requirements that your passwords must be numeric, a certain length, etc, you really need to pass them in yourself


git-svn-id: http://bigbluebutton.googlecode.com/svn/trunk@4276 af16638f-c34d-0410-8cfa-b39d5352b314