Refactor routes/index as a class MainRouter
Also removed some methods/routes that were not being used and will have to
be refactored anyway if they are needed in the future. They are: get_join,
to join the session getting the username and password from the url, and
post_upload, to upload presentations from the html5 client (this was a
specially big method).