Working through more than one proxy server
A modified version of RTMP Client is being used to communicate with the next
proxy server, this client behaves like the bbb flash client. The difference is
how the client asks the server for the stream. The bbb flash client uses only
the streamname as a parameter in play(), while the red5 client also uses
timestamp and length.