Fix off-by-one error in slide clear events
In 0.9.0, the slide clear events changed from using 1-based page numbers to
using 0-based page numbers.

Add 1 to the page number for recordings generated on a 0.9.0+ server to
fix the issue; at this point it's too late in the release to change the value,
and there's too big of a body of existing recordings out there.