Restream video after disconnect
This commit will enable the client to restream video in the case of a disconnect.

A _restream variable has been added to the VideoEventMapDelegate.as, that will dictate if the stream needs to be restream.
The setting are stored also stored so ensure that no user interaction is needed to restream the video.

VideoEventMapDelegate.as now listens to BBBEvent.RECONNECT_CONNECTION_ATTEMPT_SUCCEEDED_EVENT, and will restream the video in the case _restream is true.

_restream is set to true when the user shares his camera, and set to false when the user manually closes his camera.