fix duplicated publish video windows after reconnection #2808
The problem was that we were handling the reconnecton succeeded event to start streaming again, but this event is dispatched for every connection, not only for the video connection, and it was causing restream to be called multiple times after the reconnection, eventually duplicating the publish windows; when it occurred, the user was publishing two separate video streams.
The solution was to use the ConnectedEvent used only by the video connection, which also carries the information if that connection is due to a reconnect.
It looks like the current solution to restream after a reconnect doesn't work if the user was publishing two cameras at the same time. In that case, I think only the second camera will be published again.