HTML5 - comment out Join Video to fix DOM positioning
The reason there was so much space between `Join Audio` and `Status` was that `Join Video` was in between. Join Video was removed
to fix this positioning.

Leave Audio and Join Video icons are being hidden, which still holds the space for that component in the DOM, but does not show
it.
`Join Audio` should not be commented out because the icon is displayed at certain times.
`Join Video` can be commented out for now for UI purposes and put back when it's implemented.