bbb-conf: Enable use of domain with the word NAME
When using a domain with the word "name" in it, everything before "name" would be stripped out. bigblue.mynameisrehuel.com would be stripped to isrehuel.com. The name of my country, Suriname, has "name" in it, which presented the problem.

Changing ".*name" in the sed s/ expression in these 4 lines to ".*server_name" makes sure that "server_name" is removed, and not everything with "name". Unless these lines served a different purpose.