RaP: Don't error on pencil drawings with 0 dataPoints
In some cases, we get DRAW_END events for pencil shapes from the
html5 client that have no dataPoints. The only thing we can really
do here is detect the issue and ignore the shape.

In some cases, this may result in the shape's intermediate drawing
updates being shown, but it'll disappear when the end event happens.