Sort events when writing the events.xml file
BigBlueButton can sometimes write events out of order - this particularly
seems to affect the final RecordStatusEvent in a meeting which was ended
while recording was still running. This breaks the recording processing
scripts.

This was fixed in 2.2, but the fix doesn't merge cleanly into 2.0 because
the archiving code was rewriten due to the segment/chapter break feature.
To get the fix here, I've rewritten the store_events on 2.0 to match the
code on 2.2, that way the same sorting function can be used.

For #6035