Clarify german document upload prompt
This clarifies that a file has been _selected_ for upload, but is _not_ currently being uploaded as the former translation would suggest. Fixes #9093, thanks to @Sea444 for pointing out that this was not a quirk to work around (as I had done before) but a serious UX issue that needs fixing.