add request for audio track
This will ask for an audio track in screensharing when the user checks the `share audio` option when providing screen permission.

The entire system audio can be captured in Windows when using Chrome or Edge, but on Linux and Mac only the audio of a tab can be captured.

Due to the latest build of Edge being chromium based and the way that browser detection has been added to this project `isChrome` will report true for Chrome and Edge.