Allow user bigbluebutton to delete recordings
This patch modifies `bbb-record` allowing the user `bigbluebutton` to
delete recordings. The user has all necessary access rights, meaning
that the deletion works without a problem and the check for root does
not protect anything. The users owns the data after all. The current
check just makes things less convenient.

This is the develop variant of pull request #10841 which went into 2.2.x