Do not install Microsoft fonts in bbb-libreoffice by default
As laid out in #12035, it is illegal to install these fonts without
the end-user installing it explicitly accepting the EULA, plus, shipping
the fonts by default makes the image essentially non-free.

The packages fonts-liberation and fonts-arkpandora contain
drop-in replacements for Arial, Courier, Verdana and Times.

To install the original Microsoft fonts, the end user/administrator,
or some mechanism in bbb-install, can set SOFFICE_VARIANT=nonfree
in the environment.