implement anonymous polls option
Sometimes it is desireable to have an option for anonymous polls - think
of "How did you like my presentation?".

This patch changes three components in the HTML5 client:

* add anonymous option to poll creation form
* do not display user responses in live result for anonymous polls
* indcate whether poll is anonymous to users

In the poll creation form there is a text explaining what is going on.
It changes together with the toggle: I don't know if this is a problem
for accessability if texts change if a switch is toggled. Should this be
the case, this has to be changed.