mute/unmute state is now recognized both in translation and floor channels, so the user has full control over it.
fixed a bug that would reactivate a muted translator's mic after changing to a translation channel. a race condition when using translatorBridge.getPeerConnection() was responsible for the failed attempt to re-mute after the channel was changed. there is now retry mechanism with an increasing timeout that seems to solve the problem. a permament solution would be to introduce observables or promises to sip.js.