fix(audio): better control audio state for main and breakout rooms
The breakout room state is now retrieve directly from Breakouts collection,
and not from the previou states: this fix a race condition where
the previous state had a dirty value, leading the main audio to reconnect
in the main room while the user was actually being transfered to the breakout
room. In this situation, user had microphone active in both main and breakout
rooms.

We now let the mic state to depends only on ICE/SIP states, instead of
waiting for FreeSWITCH confirmation. This also prevents another bug/rc where
the client didn't get updated depending on the focus state of the main/breakout
room and user gets stucked at "Establishing audio connection" modal.
This also speeds up the audio connection a bit.

Closes #13636