feat(audio): add mic and listen only bridge fallback system
Introduces a (kind of) formalized system to allow using a fallback
bridge in case a microphone or listen only connection fails.

The motivation: initial field trials of the new full audio bridge.
  - Use it as fallback - whenever SIP.js/FS fails, we use the
  experimental bridge and measure whether it works when SIP.js does not.
  - Possibly increase connection success rate while also introducing
  the bridge cautiously

They can be activated via the following configuration keys:
  - public.media.audio.fallbackFullAudioBridge: defines the fallback
    mic bridge.
  - public.media.audio.fallbackListenOnlyBridge: defines the fallback
    listen only bridge.
If either of the configurations are undefined (the current default), no
fallback will be used.