fix(audio): decouple mic auth RPCs from global audio (SFU)
The SFU microphone module was using global audio authentication RPCs in
akka-apps to determine whether a mic connection is valid.

Fixes an issue where SFU microphone connections wouldn't take lock
settings in account in the permission checks.

In detail: decouple them by adding GetMicrophonePermissionReqMsg and
RespMsg. Allows flexibility on which conditionals to use in both. IMO it
is also better than renaming+adding a boolean to GetGlobalAudioPermission*
due to the fact that it gets easier to guarantee backwards compatibility
between SFU and BBB (ie SFU 2.9 keeps working on 2.4, 2.5).