refactor(audio): separate gUM and connection error handling
gUM and connection related error handlers are handled equally even
though they mean different things and should result in separate
behaviors. Connection errors are susceptible to fallback retries while
gUM errors are not. gUM errors are susceptible to non-fallback retries
(eg when an OverConstrained error pops up and we want to rollback to a
safe constraint set), but connection errors generally are not.

This adds a set of changes to make that separation clearer:
  - sfu-audio-bridge: decouple gUM factory to bubble gUM errors up
    prior to negotiation/broker creation
  - sfu-audio bridge: only attach termination/failure event handlers
    after a broker is correctly created. Creation errors should only be
    thrown by the joinAudio's return Promise
  - audio-manager: decouple gUm and connection error handling in
    joinAudio
  - audio-manager: refactor how the fallback flag behaves and is used to
    to improve readability