Revert "fix: don't clear text fake annotations"
This reverts commit 74c4c1c4ccc8f183704121a5747c769a3802bd15.
Reverting this commit because it causes fake annotations from the text
tool not to be removed. Since the text tool is the only one that
generates fake annotations(because it is the only one that is live
synced), the function to clear fake annotations is directly intended to
clear them.