Fix slides export for cluster setup
In cluster proxy setup as described in
https://docs.bigbluebutton.org/admin/clusterproxy.html you cannot assume that
the BBB Web API URL is at the same domain as the html5 client. Therefore
operators need to adjust the URL where presentation download is available.

This patch adds the ability to override the config provided by the package. This
is related to #14484