Allow webcam focus if only two webcams exist
This patch allows users to focus a webcam even if only two webcams exist. The
previous lower bound for this was three webcams.

This allows viewers to focus and therefor enlarge one webcam in situations were
two moderators share their cameras but one is the interesting one you want to
actually follow.

This fixes #15978