popup correctly shown
After a lot of try & error, I finally found a workaround. With these changes the style popup and the dots popup will be correctly shown, plus the TD-tool panel will be correctly shown on the right when the detached window that started small is expanded. Before these change, the TD-tool was shown vertically in the middle of the window and the position is minus.