Use GitHub Actions and Pages to build and deploy the docs
This commit introduces GitHub Actions to build the docs and deploy them
to GitHub Pages with the following behavior:

* The job will only be triggered for pushes to the branches 'develop' or
those starting with 'v' and changes to the 'docs'-folder in order to
avoid excessive triggering.
* Jobs that are triggered concurrently on different branches are
  canceled and only the last job will be executed.

The template for this action is taken from the original docusaurus
documentation (https://docusaurus.io/docs/deployment#deploying-to-github-pages).

Closes #16665