recording: Fix handling deskshare with delayed video
When a deskshare stream with combined audio + video starts up, it can
happen that the audio starts before the video - so the first video frame
will be some amount of time after the file start.

If there's a recording processing cut in this gap, the procesing can
crash because it can generate an output video with no video frames.

There are two parts to the fix:
  * Trim input videos with the trim filter, configured to ensure at
    least 1 output frame is generated, even if it would be after the
    end timestamp.
  * Use the tpad filter to pad the *start* of a video stream to make
    sure there's something in the gap.