recording: Skip EDL area if it is nil
The recording processing would crash if an area was present in the
layout, but was missing in the EDL entry being processed.

This doesn't happen in normal conditions, since most of the methods for
generating an EDL will result in areas being present, even if there are
no videos for that area.

But the EDL cleanup recently added can sometimes cause an EDL entry with
no areas to be processed, so add the code to handle this possibility.