recording: Increase quality of libx264 intermediate files
The encoding settings for intermediate files was using -preset veryfast
-crf 30, which resulted in very poor quality video.

After a bit of experimenting, I decided to change this to -preset
veryfast -crf 23. This results in files which are roughly twice the size
of before, but they look significantly better.

There's improvements possible at the same filesize by switching to a
slower encoding mode. But in the case of -preset medium for example,
when normalized to the same output file size, you end up using about 1.5
times as much cpu time to gain only a very small amount of video
quality.