fix: add TURN preflight check when iceTransportPolicy=relay
The built-in coturn+haproxy setup works most of times. Some folks do
have problems with it, though, and the transport policy for Firefox is
to force relay by default. That causes scenarios where Firefox fails
unconditionally with borked coturn setups and that causes more damage
than what we were trying to fix with it.

This adds a pre-flight check that is run once the application is loaded
to check whether the provided TURN server is working. The check runs at
most three times until it succeeds or fails, and results are cached for
a session. If it fails, transport policy will fall back to default
instead of trying to force relay.