build(freeswitch): apply opusfile fix from upstream as a patch
Applies commit 8e5dc5a087634145e5fc68b4e6a693d7723fe6430 from
signalwire as a patch to try to fix a crash when starting recordings.
Backtraces I've looked into point to somewhere closer to where that
commit is tweaking things and behavior seems similar. Testing is still
underway.

Revert if it doesn't work or, if it works, once we adopt a release
with it (still unreleased as of 1.10.9).