fix(timer): sidebar content when activated
Changes the behavior of activating the timer. The behavior set by this commit
is: when the timer is activated, only the moderator who triggered it will see
the sidebar content switch to the timer. This change required hanging an
observer to catch the timer activation and dispatch the sidebar content
change since timer messages are passing through akka.